Add working version of the watchdog. Need further refactoring.

+++ filters-api/src/database/migrations/20220402122949-create-transaction.js
@@ -23,7 +23,7 @@ module.exports = {
         type: Sequelize.STRING,
       },
       amount: {
-        type: Sequelize.INTEGER,
+        type: Sequelize.DECIMAL(36, 18),
       },
       filterId: {
         type: Sequelize.UUID,
